airbyte-ci/connectors/snoop/pyproject.toml
[tool.poetry]
name = "snoop"
version = "1.5.0"
description = "A package to snoop on connectors execution to generate telemetry."
authors = ["Airbyte <contact@airbyte.io>"]
readme = "README.md"
packages = [{ include = "snoop", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
requests = "^2.32.3"
pyyaml = "^6.0.1"
genson = "^1.3.0"
airbyte-protocol-models = "<1.0.0"
mitmproxy = "==10.2.4"
asyncclick = "^8.1.7.2"
semver = "^3.0.2"
google-cloud-storage = "^2.17.0"
pytest = "^8.3.1"
deepdiff = "^7.0.1"
avro = "^1.11.3"
google-cloud-pubsub = "^2.23.0"


[tool.poetry.scripts]
snoop = "snoop.cli:snoop"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
